refactor: remove status element
