+++ packages/npm/@amazeelabs/publisher-elements/src/elements.ts
@@ -1,2 +1,1 @@
-import './elements/Status';
 import './elements/Floater';
